fb_ads_library_prospecting/requirements.txt
# Facebook Ads Library Prospecting Pipeline
# Install: pip install -r requirements.txt

# Data processing
pandas>=2.0.0
openpyxl>=3.1.0

# Web requests & scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
duckduckgo-search>=4.0.0

# AI
openai>=1.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
pyyaml>=6.0.0
